docs/example patches for VMS

+++ docs/examples/ftpupload.c
@@ -5,7 +5,7 @@
  *                            | (__| |_| |  _ <| |___
  *                             \___|\___/|_| \_\_____|
  *
- * $Id: ftpupload.c,v 1.15 2008-05-22 21:20:09 danf Exp $
+ * $Id: ftpupload.c,v 1.16 2009-06-05 18:40:40 yangtse Exp $
  */
 
 #include <stdio.h>
@@ -65,7 +65,7 @@ int main(int argc, char **argv)
     printf("Couldnt open '%s': %s\n", LOCAL_FILE, strerror(errno));
     return 1;
   }
-  printf("Local file size: %ld bytes.\n", file_info.st_size);
+  printf("Local file size: %ld bytes.\n", (long)file_info.st_size);
 
   /* get a FILE * of the same file */
   hd_src = fopen(LOCAL_FILE, "rb");
